moved factionType enum and related functions to its own file
